+++ axum-extra/src/response/mod.rs
@@ -1,5 +1,8 @@
 //! Additional types for generating responses.
 
+#[cfg(any(feature = "attachment", feature = "file-stream"))]
+mod content_disposition;
+
 #[cfg(feature = "erased-json")]
 mod erased_json;